% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\name{changeparam}
\alias{changeparam}
\title{Change preprosimparametes}
\usage{
changeparam(object, contamination, param, value)
}
\arguments{
\item{object}{(preprosimparameter object)}

\item{contamination}{(character) one of the following: noise, lowvar, misvalue, irfeature}

\item{param}{(character) one of the following: cols, param, order}

\item{value}{(numeric) vector of parameter values}
}
\description{
Change preprosimparametes
}

