% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Plotting.R
\name{getpreprosimdata}
\alias{getpreprosimdata}
\title{Get the analysis results}
\usage{
getpreprosimdata(object, type = "accuracy", x, z)
}
\arguments{
\item{object}{(preprosimanalysis class object) object}

\item{type}{(character) type of data: accuracy, varimportance, outliers or xz}

\item{x}{(character) x axis contamination}

\item{z}{(character) z axis contamination}
}
\description{
Get the analysis results
}
\details{
contaminations are : noise, lowvar, misval, irfeature, classswap, classimbalance, volumedecrease, outlier
}
\examples{
## res <- preprosimrun(iris)
## getpreprosimdata(res, "accuracy")
## getpreprosimdata(res, type="xz", x="misval", z="noise")
}

