% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06Plotting.R
\name{getpreprosimdf}
\alias{getpreprosimdf}
\title{Get a contaminated data frame}
\usage{
getpreprosimdf(object, paramvector)
}
\arguments{
\item{object}{(preprosimanalysis class object) object to be plotted}

\item{paramvector}{(numeric) contamination combinations to be searched for}
}
\description{
Get a contaminated data frame
}
\examples{
## res <- preprosimrun(iris)
## df <- preprosimdf(res, c(0,0,0,0,0,0,0,0)) # returns uncontaminated original data set
}

