% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\docType{class}
\name{preprosimanalysis-class}
\alias{preprosimanalysis-class}
\title{An S4 class to represent preprosim analysis output}
\description{
An S4 class to represent preprosim analysis output
}
\section{Slots}{

\describe{
\item{\code{grid}}{(data frame) data frame consisting of combinations of preprosimparameters combinations}

\item{\code{data}}{(list) list of simulated data sets}

\item{\code{output}}{(numeric) vector of classification accuracies}

\item{\code{variableimportance}}{(data frame) data frame consisting of variable importance values}

\item{\code{outliers}}{(numeric) vector of outlier scores}
}}

