% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Classes.R
\docType{class}
\name{preprosimdata-class}
\alias{preprosimdata-class}
\title{An S4 class to represent preprosim data}
\description{
An S4 class to represent preprosim data
}
\section{Slots}{

\describe{
\item{\code{x}}{(data frame) data frame consisting of numeric features}

\item{\code{y}}{(factor) vector of class labels}

\item{\code{z}}{(boolean) safe for classification (i.e. no NAs)}
}}

