% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02Parameters.R
\docType{data}
\name{preprosimdefaultparam}
\alias{preprosimdefaultparam}
\title{Default parameters for Iris simulation runs}
\format{\preformatted{Formal class 'preprosimparameter' [package "preprosim"] with 32 slots
  ..@ noisecol              : int [1:4] 1 2 3 4
  ..@ noiseparam            : num [1:3] 0 0.1 0.2
  ..@ noiseorder            : num 1
  ..@ noisefunction         : chr "noisefunction(data, param)"
  ..@ lowvarcol             : num 2
  ..@ lowvarparam           : num [1:3] 0 0.1 0.2
  ..@ lowvarorder           : num 2
  ..@ lowvarfunction        : chr "lowvarfunction(data, param)"
  ..@ misvalcol             : int [1:4] 1 2 3 4
  ..@ misvalparam           : num [1:3] 0 0.1 0.2
  ..@ misvalorder           : num 8
  ..@ misvalfunction        : chr "misvalfunction(data, param)"
  ..@ irfeaturecol          : num 0
  ..@ irfeatureparam        : num [1:3] 0 0.1 0.2
  ..@ irfeatureorder        : num 3
  ..@ irfeaturefunction     : chr "irfeaturefunction(data, param)"
  ..@ classswapcol          : num 0
  ..@ classswapparam        : num [1:3] 0 0.1 0.2
  ..@ classswaporder        : num 4
  ..@ classswapfunction     : chr "classswapfunction(data, param)"
  ..@ classimbalancecol     : num 0
  ..@ classimbalanceparam   : num [1:3] 0 0.1 0.2
  ..@ classimbalanceorder   : num 5
  ..@ classimbalancefunction: chr "classimbalancefunction(data, param)"
  ..@ volumedecreasecol     : num 0
  ..@ volumedecreaseparam   : num [1:3] 0 0.1 0.2
  ..@ volumedecreaseorder   : num 6
  ..@ volumedecreasefunction: chr "volumedecreasefunction(data, param)"
  ..@ outliercol            : num 3
  ..@ outlierparam          : num [1:3] 0 0.1 0.2
  ..@ outlierorder          : num 7
  ..@ outlierfunction       : chr "outlierfunction(data, param)"
}}
\usage{
preprosimdefaultparam
}
\description{
Defaultparameters include eight data contaminations: noise, low variance, missing values,
irrelevant features, class label swap, class imbalance, volume decrease and outliers.
}
\details{
For each contamination there is column (integer vector) the contamination is applied to
in the Iris dataset, parameters for contamination (integer vector) corresponding to zero,
ten percent and twenty percent contamination intensity and order in which contaminations
are applied.

There are total 6561 combinations of parameters in the preprosimdefaultparameters.
}
\keyword{datasets}

