% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\docType{class}
\name{preprosimparameter-class}
\alias{preprosimparameter-class}
\title{An S4 class to represent preprosimparameter}
\description{
An S4 class to represent preprosimparameter
}
\section{Slots}{

\describe{
\item{\code{noisecol}}{(numeric)}

\item{\code{noiseparam}}{(numeric)}

\item{\code{noiseorder}}{(numeric)}

\item{\code{noisefunction}}{(character)}

\item{\code{lowvarcol}}{(numeric)}

\item{\code{lowvarparam}}{(numeric)}

\item{\code{lowvarorder}}{(numeric)}

\item{\code{lowvarfunction}}{(character)}

\item{\code{misvalcol}}{(numeric)}

\item{\code{misvalparam}}{(numeric)}

\item{\code{misvalorder}}{(numeric)}

\item{\code{misvalfunction}}{(character)}

\item{\code{irfeaturecol}}{(numeric)}

\item{\code{irfeatureparam}}{(numeric)}

\item{\code{irfeatureorder}}{(numeric)}

\item{\code{irfeaturefunction}}{(character)}

\item{\code{classswapparam}}{(numeric)}

\item{\code{classswaporder}}{(numeric)}

\item{\code{classswapfunction}}{(character)}

\item{\code{classimbalanceparam}}{(numeric)}

\item{\code{classimbalanceorder}}{(numeric)}

\item{\code{classimbalancefunction}}{(character)}

\item{\code{volumedecreaseparam}}{(numeric)}

\item{\code{volumedecreaseorder}}{(numeric)}

\item{\code{volumedecreasefunction}}{(character)}

\item{\code{outlierparam}}{(numeric)}

\item{\code{outlierorder}}{(numeric)}

\item{\code{outlierfunction}}{(character)}
}}

