% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05Plotting.R
\name{preprosimplot}
\alias{preprosimplot}
\title{Plot data quality simulation results}
\usage{
preprosimplot(object, type, xaxis)
}
\arguments{
\item{object}{(preprosimanalysis class object) object to be plotted}

\item{type}{(character) type of plot: accuracy, varimportance, outliers}
}
\description{
Plot data quality simulation results
}

