% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\name{changeparam}
\alias{changeparam}
\title{Change preprosimparametes}
\usage{
changeparam(object, contamination, param, value)
}
\arguments{
\item{object}{(preprosimparameter object)}

\item{contamination}{(character) one of the following: noise, lowvar, misvalue, irfeature, classswap, classimbalance, volumedecrease, outlier}

\item{param}{(character) one of the following: cols, param, order}

\item{value}{(numeric) scalar (for order) or vector (for cols and param) of parameter values}
}
\description{
Change preprosimparametes
}
\details{
The order of contaminations (cols parameter) must be between 1 and 8, and no two contaminations can have the same order
}
\examples{
## res <- newparam()
## res <- changeparam(res, "noise", "cols", value=1)
## res <- changeparam(res, "noise", "param", value=c(0,0.1))
## res <- changeparam(res, "noise", "order", value=1)
}

