% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\name{newparam}
\alias{newparam}
\title{Create new preprosimparameter object}
\usage{
newparam()
}
\value{
preprosimparameter class object
}
\description{
Create new preprosimparameter object
}
\details{
all columns, parameters and order are set to zero
}
\examples{
## res <- newparam()
}

