% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Classes.R
\name{newparam}
\alias{newparam}
\title{Create new preprosimparameter object}
\usage{
newparam(dataframe, type = "default", x, z)
}
\arguments{
\item{dataframe}{(data frame) original data to be used in simulations}

\item{type}{(character) creation type: empty, default or custom, defaults to "default"}

\item{x}{(character) primary contamination of interest such as "misval"}

\item{z}{(character) secondary contamination of interest such as "noise"}
}
\value{
preprosimparameter class object
}
\description{
Create new preprosimparameter object
}
\details{
empty creates a preprosimparamater object with empty params; default
}
\examples{
## res <- newparam(iris)
## res <- newparam(iris, "empty")
## res <- newparam(iris, "custom", "misval", "noise")
}

