% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02Parameters.R
\docType{data}
\name{preprosimdefaultparam}
\alias{preprosimdefaultparam}
\title{preprosimdefaultparam}
\format{\preformatted{Formal class 'preprosimparameter' [package "preprosim"] with 32 slots
  ..@ noisecol              : num 1
  ..@ noiseparam            : num [1:2] 0 0.5
  ..@ noiseorder            : num 1
  ..@ noisefunction         : chr "noisefunction(data, param)"
  ..@ lowvarcol             : num 2
  ..@ lowvarparam           : num [1:2] 0 0.1
  ..@ lowvarorder           : num 2
  ..@ lowvarfunction        : chr "lowvarfunction(data, param)"
  ..@ misvalcol             : int [1:4] 1 2 3 4
  ..@ misvalparam           : num [1:5] 0 0.1 0.2 0.3 0.4
  ..@ misvalorder           : num 8
  ..@ misvalfunction        : chr "misvalfunction(data, param)"
  ..@ irfeaturecol          : num 4
  ..@ irfeatureparam        : num [1:2] 0 0.1
  ..@ irfeatureorder        : num 3
  ..@ irfeaturefunction     : chr "irfeaturefunction(data, param)"
  ..@ classswapcol          : num 4
  ..@ classswapparam        : num [1:2] 0 0.1
  ..@ classswaporder        : num 4
  ..@ classswapfunction     : chr "classswapfunction(data, param)"
  ..@ classimbalancecol     : num 4
  ..@ classimbalanceparam   : num [1:2] 0 0.1
  ..@ classimbalanceorder   : num 5
  ..@ classimbalancefunction: chr "classimbalancefunction(data, param)"
  ..@ volumedecreasecol     : num 4
  ..@ volumedecreaseparam   : num [1:2] 0 0.1
  ..@ volumedecreaseorder   : num 6
  ..@ volumedecreasefunction: chr "volumedecreasefunction(data, param)"
  ..@ outliercol            : num 4
  ..@ outlierparam          : num [1:2] 0 0.1
  ..@ outlierorder          : num 7
  ..@ outlierfunction       : chr "outlierfunction(data, param)"
}}
\usage{
preprosimdefaultparam
}
\description{
defaultparameters include eight data contaminations each having three parameter levels
}
\keyword{datasets}

