% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/07Plotting.R
\name{preprosimplot}
\alias{preprosimplot}
\title{Plot data quality simulation results}
\usage{
preprosimplot(object, type = "accuracy", x, z)
}
\arguments{
\item{object}{(preprosimanalysis class object) object to be plotted}

\item{type}{(character) type of plot: accuracy, varimportance, outliers or xz; defaults to accuracy}

\item{x}{(character) x axis contamination}

\item{z}{(character) z axis contamination plotted as panels}
}
\description{
Plot data quality simulation results
}
\details{
contaminations are : noise, lowvar, misval, irfeature, classswap, classimbalance, volumedecrease, outlier
}
\examples{
## res <- preprosimrun(iris)
## preprosimplot(res)
## preprosimplot(res, type="xz", x="misval", z="noise")
}

