% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04Preprosim.R
\name{preprosimrun}
\alias{preprosimrun}
\title{Run data quality simulation}
\usage{
preprosimrun(data, param = preprosimdefaultparam, seed = 1,
  holdoutrounds = 10)
}
\arguments{
\item{data}{(data frame) one factor columns for class labels, other columns numeric, no missing values}

\item{param}{(preprosimparameter object) simulation parameters, defaults to preprosimdefaultparameters}

\item{seed}{(integer) seed to be used for reproducible results}

\item{holdoutrounds}{(integer) number of holdout rounds,default to 10}
}
\value{
preprosimanalysis class object
}
\description{
Run data quality simulation
}

