% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/04Preprosim.R
\name{preprosimrun}
\alias{preprosimrun}
\title{Run data quality simulation}
\usage{
preprosimrun(data, param = newparam(data, "default"), seed = 1,
  caretmodel = "gbm", holdoutrounds = 10, cores = 1)
}
\arguments{
\item{data}{(data frame) one factor columns for class labels, other columns numeric, no missing values}

\item{param}{(preprosimparameter object) simulation parameters, defaults to parameters set automatically for data.}

\item{seed}{(integer) seed to be used for reproducible results, defaults to 1}

\item{caretmodel}{(character) a model from package Caret, defaults to gbm (gbm must be installed before preprosimrun)}

\item{holdoutrounds}{(integer) number of holdout rounds, defaults to 10}

\item{cores}{(integer) number of cores used in parallel processing, defaults to 1}
}
\value{
preprosimanalysis class object
}
\description{
Run data quality simulation
}
\examples{
## res <- preprosimrun(iris)
## res1 <- preprosimrun(iris, param=newparam(iris, "custom", x="misval", z="noise"), cores=2)
## res2 <- preprosimrun(iris, caretmodel="rpart")
}

